Extract matching logic to match_mcsc_to_rc.py
Move archive-to-show matching functions from generate-show-cache.py
to a dedicated module for better separation of concerns:

- normalize_text() - text normalization for comparison
- extract_date_from_title() - parses 10+ date formats
- calculate_match_score() - scoring function
- match_archives_to_shows() - main matching orchestrator

The should_exclude_show() function stays in the main script and is
passed as a callback to keep config-dependent logic together.
